LeetCode Problem - 37
Problem Link - https://leetcode.com/problems/sudoku-solver/

Approach:- 
Using backtracking this problem can be solved. Find the first blank cell and try to put each value from 1 to 9. If a value is valid for the current cell (i.e. not present in that particular row, column and 3x3 cell) then put that value in the cell and call the function recursively to fill rest of the blank cells.